+++ eform-client/wdio-headless-plugin-step2.conf.js
@@ -11,7 +11,11 @@ exports.config = {
   //
   specs: [
     'e2e/Tests/items-planning-settings/application-settings.plugins-page.spec.ts',
-    'e2e/Tests/items-planning-general/*.spec.ts',
+    'e2e/Tests/items-planning-general/items-planning.add.spec.ts',
+    'e2e/Tests/items-planning-general/items-planning.edit.spec.ts',
+    'e2e/Tests/items-planning-general/items-planning.delete.spec.ts',
+    'e2e/Tests/items-planning-general/items-planning.settings.spec.ts',
+    'e2e/Tests/items-planning-general/items-planning.sorting.spec.ts'
   ],
   suites: {
     settings: [
