+++ eform-client/wdio-plugin-step2.conf.js
@@ -10,8 +10,12 @@ exports.config = {
     // directory is where your package.json resides, so `wdio` will be called from there.
     //
   specs: [
-    'e2e/Tests/items-planning-settings/application-settings.plugins-page.spec.ts',
-    'e2e/Tests/items-planning-general/*.spec.ts',
+      'e2e/Tests/items-planning-settings/application-settings.plugins-page.spec.ts',
+      'e2e/Tests/items-planning-general/items-planning.add.spec.ts',
+      'e2e/Tests/items-planning-general/items-planning.edit.spec.ts',
+      'e2e/Tests/items-planning-general/items-planning.delete.spec.ts',
+      'e2e/Tests/items-planning-general/items-planning.settings.spec.ts',
+      'e2e/Tests/items-planning-general/items-planning.sorting.spec.ts'
   ],
     suites: {
         settings: [
